PR review: minor fix to account for comments
- move files
- by default serialization in py bindings, remove VAR
