+++ CMakeLists.txt
@@ -147,7 +147,6 @@ if(BUILD_TESTING OR BUILD_PYTHON_INTERFACE)
   set(cereal ${cereal_dir}/README.md)
   find_package(Git REQUIRED)
   if(NOT EXISTS ${cereal})
-    message("Initialize cereal submodule.")
     execute_process(
       COMMAND ${GIT_EXECUTABLE} submodule update --init ${cereal_dir}
       WORKING_DIRECTORY ${cereal_dir} COMMAND_ERROR_IS_FATAL ANY)
