+++ include/proxsuite/serialization/results.hpp
