+++ include/proxsuite/serialization/settings.hpp
